Create an OrdersReportTable component
